refactor: Extract standings utilities to separate module
- Move standings API logic from useStandings.ts to dedicated standingsUtils.ts
- Update usePrefetchData.ts to import from new utilities module
- Add ESPN API domains to manifest host_permissions for proper API access
- Update CSP to allow connections to ESPN API endpoints
- Improve code organization and maintainability
